requirements/PROD.txt
python-telegram-bot~=20.8
pydantic~=2.4.2
pydantic-settings
httpx~=0.26.0